Use 32-bit integers to represent positions and scaling when rendering rectangles and textures

+++ src/renderer.c
@@ -86,7 +86,7 @@ void renderer_init() {
   tile_shader = shader_compile("resource/shader/tile");
 }
 
-void render_texture_region(struct Texture texture, float x, float y, float z, float w, float h, float angle, i32 x_offset, i32 y_offset, i32 x_range, i32 y_range) {
+void render_texture_region(struct Texture texture, i32 x, i32 y, float z, i32 w, i32 h, float angle, i32 x_offset, i32 y_offset, i32 x_range, i32 y_range) {
   const struct Texture_program* program = &texture_program;
   glUseProgram(program->handle);
 
@@ -179,7 +179,7 @@ void render_text(struct Texture font_texture, float x, float y, float z, float w
   glBindVertexArray(0);
 }
 
-void render_rect(float x, float y, float z, float w, float h, float r, float g, float b, float a, float angle, float thickness) {
+void render_rect(i32 x, i32 y, float z, i32 w, i32 h, float r, float g, float b, float a, float angle, float thickness) {
   const u32 program = rect_shader;
   glUseProgram(program);
 
@@ -204,7 +204,7 @@ void render_rect(float x, float y, float z, float w, float h, float r, float g, 
   glBindVertexArray(0);
 }
 
-void render_filled_rectangle(float x, float y, float z, float w, float h, float r, float g, float b, float a, float border_r, float border_g, float border_b, float border_a, float angle, float thickness) {
+void render_filled_rectangle(i32 x, i32 y, float z, i32 w, i32 h, float r, float g, float b, float a, float border_r, float border_g, float border_b, float border_a, float angle, float thickness) {
   const u32 program = filled_rect_shader;
   glUseProgram(program);
